Move logged in state to logged in router
